+++ src/cpp/hydra.h
@@ -10,14 +10,20 @@
 #define HYDRA_H
 
 #include<sstream>
+#include<functional>
 
 #include<error_tools.h>
+#define USE_EIGEN
+#include<vector_tools.h>
 #include<abaqus_tools.h>
 
 namespace hydra{
 
+    // forward class definitions
+    class hydraBase;
+
     namespace unit_test{
-        class hydraTester;
+        class hydraBaseTester;
     }
 
     constexpr const char* str_end(const char *str) {
@@ -59,9 +65,88 @@ namespace hydra{
     typedef std::vector< floatType > floatVector; //!< Define a vector of floats
     typedef std::vector< std::vector< floatType > > floatMatrix; //!< Define a matrix of floats
 
-    class hydra{
+    typedef void ( hydraBase::*hydraBaseFxn )( ); //!< Typedef for passing pointers to hydraBase functions
+
+    class dataBase{
+
+        public:
+
+            virtual void clear( ){
+                /*!
+                 * The function to erase the current values stored
+                 */
+
+                ERROR_TOOLS_CATCH( throw std::runtime_error( "clear not implemented!" ) );
+
+            }
+
+    };
+
+    template < typename T >
+    class dataStorage : public dataBase{
+
+        public:
+
+            bool first = false; //!The flag for whether the data has been stored
+
+            T second; //!The stored data
+
+            dataStorage( ){ };
+
+            dataStorage( const bool &_first, const T &_second ) : first( _first ), second( _second ) { }
+
+            virtual void clear( ){
+                /*!
+                 * The function to erase the current values stored by setting first to false and clearing second
+                 */
+
+                first = false;
+
+                second.clear( );
+
+            }
+
+    };
+
+    template <>
+    inline void dataStorage< int >::clear( ){
+                /*!
+                 * The function to erase the current values stored by setting first to false and second to zero
+                 */
+
+        first = false;
+
+        second = 0;
+
+    }
+
+    template <>
+    inline void dataStorage< unsigned int >::clear( ){
+                /*!
+                 * The function to erase the current values stored by setting first to false and second to zero
+                 */
+
+        first = false;
+
+        second = 0;
+
+    }
+
+    template <>
+    inline void dataStorage< floatType >::clear( ){
+                /*!
+                 * The function to erase the current values stored by setting first to false and second to zero
+                 */
+
+        first = false;
+
+        second = 0;
+
+    }
+
+    class hydraBase{
         /*!
-         * Hydra: A base class which can be used to construct finite deformation material models.
+         * hydraBase: A base class which can be used to construct finite deformation material models.
          * 
          * The hydra class seeks to provide utilities for the construction of finite deformation constitutive models
          * more rapidly than would be possible previously. The user can define as many different configurations as desired
@@ -72,43 +157,130 @@ namespace hydra{
 
         public:
 
+            // Constructors
+            //! Default constructor for hydraBase
+            hydraBase( ){ }
+
+            //! Main constructor for objects of type hydraBase. Sets all quantities required for most solves.
+            hydraBase( const floatType &time, const floatType &deltaTime,
+                       const floatType &temperature, const floatType &previousTemperature,
+                       const floatVector &deformationGradient, const floatVector &previousDeformationGradient,
+                       const floatVector &previousStateVariables, const floatVector &parameters,
+                       const unsigned int numConfigurations, const unsigned int numNonLinearSolveStateVariables,
+                       const unsigned int dimension=3 ) : _time( time ), _deltaTime( deltaTime ),
+                                                          _temperature( temperature ), _previousTemperature( previousTemperature ),
+                                                          _deformationGradient( deformationGradient ),
+                                                          _previousDeformationGradient( previousDeformationGradient ),
+                                                          _previousStateVariables( previousStateVariables ),
+                                                          _parameters( parameters ),
+                                                          _numConfigurations( numConfigurations ),
+                                                          _numNonLinearSolveStateVariables( numNonLinearSolveStateVariables ),
+                                                          _dimension( dimension ){ }
+
+            // User defined functions
+
+            // Setter functions
+
             // Getter functions
+            //! Get a reference to the current time
             const floatType* getTime( ){ return &_time; }
 
+            //! Get a reference to the change in time
             const floatType* getDeltaTime( ){ return &_deltaTime; }
 
+            //! Get a reference to the current temperature
             const floatType* getTemperature( ){ return &_temperature; };
 
+            //! Get a reference to the previous temperature
             const floatType* getPreviousTemperature( ){ return &_previousTemperature; };
 
+            //! Get a reference to the deformation gradient
             const floatVector* getDeformationGradient( ){ return &_deformationGradient; }
 
+            //! Get a reference to the previous deformation gradient
             const floatVector* getPreviousDeformationGradient( ){ return &_previousDeformationGradient; }
 
+            //! Get a reference to the previous values of the state variables
             const floatVector* getPreviousStateVariables( ){ return &_previousStateVariables; }
 
+            //! Get a reference to the model parameters
             const floatVector* getParameters( ){ return &_parameters; }
+
+            //! Get a reference to the number of configurations
+            const unsigned int* getNumConfigurations( ){ return &_numConfigurations; }
+
+            //! Get a reference to the number of state variables involved in the non-linear solve
+            const unsigned int* getNumNonLinearSolveStateVariables( ){ return &_numNonLinearSolveStateVariables; }
+
+            //! Get a reference to the dimension
+            const unsigned int* getDimension( ){ return &_dimension; }
+
+            //! Get a reference to the configurations
+            const floatMatrix* getConfigurations( ){ return &_configurations.second; }
+
+            //! Get a reference to the previous configurations
+            const floatMatrix* getPreviousConfigurations( ){ return &_previousConfigurations.second; }
+
+            //! Get a reference to the inverse configurations
+            const floatMatrix* getInverseConfigurations( ){ return &_inverseConfigurations.second; }
+
+            //! Get a reference to the previous inverse configurations
+            const floatMatrix* getPreviousInverseConfigurations( ){ return &_previousInverseConfigurations.second; }
+
+            //! Get a reference to the state variables used in the unknown vector of the non-linear solve
+            const floatVector* getNonLinearSolveStateVariables( ){ return &_nonLinearSolveStateVariables.second; }
+
+            //! Get a reference to the previous values of the state variables used in the unknown vector of the non-linear solve
+            const floatVector* getPreviousNonLinearSolveStateVariables( ){ return &_previousNonLinearSolveStateVariables.second; }
+
+            //! Get a reference to the state variables not used in the unknown vector for the non-linear solve
+            const floatVector* getAdditionalStateVariables( ){ return &_additionalStateVariables.second; }
+
+            //! Get a reference to the previous values of the state variables not used in the unknown vector for the non-linear solve
+            const floatVector* getPreviousAdditionalStateVariables( ){ return &_previousAdditionalStateVariables.second; }
 
         private:
 
             // Friend classes
-            friend class unit_test::hydraTester;
+            friend class unit_test::hydraBaseTester; //!< Friend class which allows modification of private variables. ONLY TO BE USED FOR TESTING!
 
-            floatType _time;
+            floatType _time; //!< The current time
 
-            floatType _deltaTime;
+            floatType _deltaTime; //!< The change in time
 
-            floatType _temperature;
+            floatType _temperature; //!< The current temperature
 
-            floatType _previousTemperature;
+            floatType _previousTemperature; //!< The previous temperature
 
-            floatVector _deformationGradient;
+            floatVector _deformationGradient; //!< The current deformation gradient
 
-            floatVector _previousDeformationGradient;
+            floatVector _previousDeformationGradient; //!< The previous deformation gradient
 
-            floatVector _previousStateVariables;
+            floatVector _previousStateVariables; //!< The previous state variables
 
-            floatVector _parameters;
+            floatVector _parameters; //!< The model parameters
+
+            unsigned int _numConfigurations; //!< The number of configurations
+
+            unsigned int _numNonLinearSolveStateVariables; //!< The number of state variables which will be solved in the Newton-Raphson loop
+            unsigned int _dimension; //!< The spatial dimension of the problem
+
+            dataStorage< floatMatrix > _configurations; //!< The current values of the configurations
+
+            dataStorage< floatMatrix > _previousConfigurations; //!< The previous values of the configurations
+
+            dataStorage< floatMatrix > _inverseConfigurations; //!< The inverses of the configurations
+
+            dataStorage< floatMatrix > _previousInverseConfigurations; //!< The inverses of the previous configurations
+
+            dataStorage< floatVector > _nonLinearSolveStateVariables; //!< The current values of the state variables involved in the non-linear solve
+            dataStorage< floatVector > _previousNonLinearSolveStateVariables; //!< The previous values of the state variables involved in the non-linear solve
+
+            dataStorage< floatVector > _additionalStateVariables; //!< The current values of the additional state variables
+
+            dataStorage< floatVector > _previousAdditionalStateVariables; //!< The previous values of the additional state variables
+
+            virtual void decomposeStateVariableVector( );
 
     };