+++ src/cpp/tests/test_hydra.cpp
@@ -14,6 +14,9 @@
 
 typedef errorTools::Node errorNode; //!< Redefinition for the error node
 typedef errorNode* errorOut; //!< Redefinition for a pointer to the error node
+typedef hydra::floatType floatType; //!< Redefinition of the floating point type
+typedef hydra::floatVector floatVector; //!< Redefinition of the vector of floating points type
+typedef hydra::floatMatrix floatMatrix; //!< Redefinition of the matrix of floating points type
 
 struct cout_redirect{
     cout_redirect( std::streambuf * new_buffer)
@@ -32,56 +35,135 @@ namespace hydra{
 
     namespace unit_test{
 
-        class hydraTester{
+        class hydraBaseTester{
 
             public:
 
-                static void checkTime( hydra &hydra ){
+                static void checkTime( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._time == hydra.getTime( ) );
     
                 }
     
-                static void checkDeltaTime( hydra &hydra ){
+                static void checkDeltaTime( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._deltaTime == hydra.getDeltaTime( ) );
     
                 }
     
-                static void checkTemperature( hydra &hydra ){
+                static void checkTemperature( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._temperature == hydra.getTemperature( ) );
     
                 }
     
-                static void checkPreviousTemperature( hydra &hydra ){
+                static void checkPreviousTemperature( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._previousTemperature == hydra.getPreviousTemperature( ) );
     
                 }
     
-                static void checkDeformationGradient( hydra &hydra ){
+                static void checkDeformationGradient( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._deformationGradient == hydra.getDeformationGradient( ) );
     
                 }
     
-                static void checkPreviousDeformationGradient( hydra &hydra ){
+                static void checkPreviousDeformationGradient( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._previousDeformationGradient == hydra.getPreviousDeformationGradient( ) );
     
                 }
     
-                static void checkPreviousStateVariables( hydra &hydra ){
+                static void checkPreviousStateVariables( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._previousStateVariables == hydra.getPreviousStateVariables( ) );
     
                 }
     
-                static void checkParameters( hydra &hydra ){
+                static void checkParameters( hydraBase &hydra ){
     
                     BOOST_CHECK( &hydra._parameters == hydra.getParameters( ) );
     
+                }
+
+                static void checkNumConfigurations( hydraBase &hydra ){
+    
+                    BOOST_CHECK( &hydra._numConfigurations == hydra.getNumConfigurations( ) );
+    
+                }
+
+                static void checkNumNonLinearSolveStateVariables( hydraBase &hydra ){
+    
+                    BOOST_CHECK( &hydra._numNonLinearSolveStateVariables == hydra.getNumNonLinearSolveStateVariables( ) );
+    
+                }
+
+                static void checkDimension( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._dimension == hydra.getDimension( ) );
+
+                }
+
+                static void checkConfigurations( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._configurations.second == hydra.getConfigurations( ) );
+
+                }
+
+                static void checkPreviousConfigurations( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._previousConfigurations.second == hydra.getPreviousConfigurations( ) );
+
+                }
+
+                static void checkInverseConfigurations( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._inverseConfigurations.second == hydra.getInverseConfigurations( ) );
+
+                }
+
+                static void checkPreviousInverseConfigurations( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._previousInverseConfigurations.second == hydra.getPreviousInverseConfigurations( ) );
+
+                }
+
+                static void checkNonLinearSolveStateVariables( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._nonLinearSolveStateVariables.second == hydra.getNonLinearSolveStateVariables( ) );
+
+                }
+
+                static void checkPreviousNonLinearSolveStateVariables( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._previousNonLinearSolveStateVariables.second == hydra.getPreviousNonLinearSolveStateVariables( ) );
+
+                }
+
+                static void checkAdditionalStateVariables( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._additionalStateVariables.second == hydra.getAdditionalStateVariables( ) );
+
+                }
+
+                static void checkPreviousAdditionalStateVariables( hydraBase &hydra ){
+
+                    BOOST_CHECK( &hydra._previousAdditionalStateVariables.second == hydra.getPreviousAdditionalStateVariables( ) );
+
+                }
+
+                static void decomposeStateVariableVector( hydraBase &hydra ){
+
+                    try{
+                        hydra.decomposeStateVariableVector( );
+                    }
+                    catch(std::exception &e){
+                        errorTools::printNestedExceptions( e );
+                    }
+
+                    ERROR_TOOLS_CATCH( hydra.decomposeStateVariableVector( ) );
+
                 }
 
         };
@@ -245,64 +327,256 @@ BOOST_AUTO_TEST_CASE( testAbaqusInterface ){
 
 BOOST_AUTO_TEST_CASE( test_hydra_getTime ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkTime( hydra );
+    hydra::unit_test::hydraBaseTester::checkTime( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getDeltaTime ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkDeltaTime( hydra );
+    hydra::unit_test::hydraBaseTester::checkDeltaTime( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getTemperature ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkTemperature( hydra );
+    hydra::unit_test::hydraBaseTester::checkTemperature( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getPreviousTemperature ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkPreviousTemperature( hydra );
+    hydra::unit_test::hydraBaseTester::checkPreviousTemperature( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getDeformationGradient ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkDeformationGradient( hydra );
+    hydra::unit_test::hydraBaseTester::checkDeformationGradient( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getPreviousDeformationGradient ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkPreviousDeformationGradient( hydra );
+    hydra::unit_test::hydraBaseTester::checkPreviousDeformationGradient( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getPreviousStateVariables ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkPreviousStateVariables( hydra );
+    hydra::unit_test::hydraBaseTester::checkPreviousStateVariables( hydra );
 
 }
 
 BOOST_AUTO_TEST_CASE( test_hydra_getParameters ){
 
-    hydra::hydra hydra;
+    hydra::hydraBase hydra;
 
-    hydra::unit_test::hydraTester::checkParameters( hydra );
+    hydra::unit_test::hydraBaseTester::checkParameters( hydra );
 
 }
+
+BOOST_AUTO_TEST_CASE( test_hydra_getNumConfigurations ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkNumConfigurations( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getNumNonLinearSolveStateVariables ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkNumNonLinearSolveStateVariables( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getConfigurations ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkConfigurations( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getPreviousConfigurations ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkPreviousConfigurations( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getInverseConfigurations ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkInverseConfigurations( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getPreviousInverseConfigurations ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkPreviousInverseConfigurations( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getNonLinearSolveStateVariables ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkNonLinearSolveStateVariables( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getPreviousNonLinearSolveStateVariables ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkPreviousNonLinearSolveStateVariables( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getAdditionalStateVariables ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkAdditionalStateVariables( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getPreviousAdditionalStateVariables ){
+
+    hydra::hydraBase hydra;
+
+    hydra::unit_test::hydraBaseTester::checkPreviousAdditionalStateVariables( hydra );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_decomposeStateVariableVector ){
+
+    floatType time = 1.1;
+
+    floatType deltaTime = 2.2;
+
+    floatType temperature = 5.3;
+
+    floatType previousTemperature = 23.4;
+
+    floatVector deformationGradient = { 0.39293837, -0.42772133, -0.54629709,
+                                        0.10262954,  0.43893794, -0.15378708,
+                                        0.9615284 ,  0.36965948, -0.0381362 };
+
+    floatVector previousDeformationGradient = { -0.21576496, -0.31364397,  0.45809941,
+                                                -0.12285551, -0.88064421, -0.20391149,
+                                                 0.47599081, -0.63501654, -0.64909649 };
+
+    floatVector previousStateVariables = { 0.53155137, 0.53182759, 0.63440096, 0.84943179, 0.72445532,
+                                           0.61102351, 0.72244338, 0.32295891, 0.36178866, 0.22826323,
+                                           0.29371405, 0.63097612, 0.09210494, 0.43370117, 0.43086276,
+                                           0.4936851 , 0.42583029, 0.31226122, 0.42635131, 0.89338916,
+                                           0.94416002, 0.50183668, 0.62395295, 0.1156184 , 0.31728548,
+                                           0.41482621, 0.86630916, 0.25045537, 0.48303426, 0.98555979,
+                                           0.51948512, 0.61289453, 0.12062867, 0.8263408 , 0.60306013,
+                                           0.54506801, 0.34276383, 0.30412079 }; 
+
+    floatMatrix configurationsAnswer = {
+                                           { 1.05936416, -0.30634264, -0.86204929,
+                                             0.03274673,  0.17917379, -0.22403642,
+                                             1.24895144, -0.21368066, -1.05360316 },
+                                           { 1.53155137,  0.53182759,  0.63440096,
+                                             0.84943179,  1.72445532,  0.61102351,
+                                             0.72244338,  0.32295891,  1.36178866 },
+                                           { 1.22826323,  0.29371405,  0.63097612,
+                                             0.09210494,  1.43370117,  0.43086276,
+                                             0.4936851 ,  0.42583029,  1.31226122 },
+                                           { 1.42635131,  0.89338916,  0.94416002,
+                                             0.50183668,  1.62395295,  0.1156184 ,
+                                             0.31728548,  0.41482621,  1.86630916 }
+                                       };
+
+    floatMatrix previousConfigurationsAnswer = {
+                                                   { -0.41803693, -0.10444357,  0.58891991,
+                                                      0.33131016, -0.34276201, -0.04604603,
+                                                      1.32949612, -0.42712653, -1.03631144 },
+                                                   {  1.53155137,  0.53182759,  0.63440096,
+                                                      0.84943179,  1.72445532,  0.61102351,
+                                                      0.72244338,  0.32295891,  1.36178866 },
+                                                   {  1.22826323,  0.29371405,  0.63097612,
+                                                      0.09210494,  1.43370117,  0.43086276,
+                                                      0.4936851 ,  0.42583029,  1.31226122 },
+                                                   {  1.42635131,  0.89338916,  0.94416002,
+                                                      0.50183668,  1.62395295,  0.1156184 ,
+                                                      0.31728548,  0.41482621,  1.86630916 }
+                                               };
+
+    floatMatrix inverseConfigurationsAnswer( 4 );
+
+    floatMatrix previousInverseConfigurationsAnswer( 4 );
+
+    for ( unsigned int i = 0; i < 4; i++ ){
+
+        inverseConfigurationsAnswer[ i ]         = vectorTools::inverse( configurationsAnswer[ i ], 3, 3 );
+        previousInverseConfigurationsAnswer[ i ] = vectorTools::inverse( previousConfigurationsAnswer[ i ], 3, 3 );
+
+    }
+
+    floatVector nonLinearSolveStateVariablesAnswer = { 0.25045537, 0.48303426, 0.98555979,
+                                                       0.51948512, 0.61289453 };
+
+    floatVector previousNonLinearSolveStateVariablesAnswer = { 0.25045537, 0.48303426, 0.98555979,
+                                                               0.51948512, 0.61289453 };
+
+    floatVector additionalStateVariablesAnswer = { 0.12062867, 0.8263408 , 0.60306013,
+                                                   0.54506801, 0.34276383, 0.30412079  };
+
+    floatVector previousAdditionalStateVariablesAnswer = { 0.12062867, 0.8263408 , 0.60306013,
+                                                           0.54506801, 0.34276383, 0.30412079  };
+
+    floatVector parameters = { 1, 2, 3, 4, 5 };
+
+    unsigned int numConfigurations = 4;
+
+    unsigned int numNonLinearSolveStateVariables = 5;
+
+    unsigned int dimension = 3;
+
+    hydra::hydraBase hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                            previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+    hydra::unit_test::hydraBaseTester::decomposeStateVariableVector( hydra );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( configurationsAnswer, *hydra.getConfigurations( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( previousConfigurationsAnswer, *hydra.getPreviousConfigurations( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( inverseConfigurationsAnswer, *hydra.getInverseConfigurations( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( previousInverseConfigurationsAnswer, *hydra.getPreviousInverseConfigurations( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( nonLinearSolveStateVariablesAnswer, *hydra.getNonLinearSolveStateVariables( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( previousNonLinearSolveStateVariablesAnswer, *hydra.getPreviousNonLinearSolveStateVariables( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( additionalStateVariablesAnswer, *hydra.getAdditionalStateVariables( ) ) );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( previousAdditionalStateVariablesAnswer, *hydra.getPreviousAdditionalStateVariables( ) ) );
+
+}
+