+++ docs/sphinx/changelog.rst
@@ -15,3 +15,5 @@ Internal Changes
 - Added getters for the base quantities (:merge:`2`). By `Nathan Miller`_.
 - Added additional libraries required for the project to update the environment (:merge:`4`). By `Nathan Miller`_.
 - Updated the environment.txt file to reflect the new recipe (:merge:`5`). By `Nathan Miller`_.
+- Added the decomposition of the incoming state variable vector into the configurations, state variables
+  in the non-linear solve, and additional state variables (:merge:`3`). By `Nathan Miller`_.